requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
Pillow>=9.5.0

# Machine Learning
scikit-learn>=1.3.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0

# Optional: For better performance
# opencv-python>=4.8.0  # Uncomment if needed for image processing

